changed PakLib's FTell return to long to match base ftell and fixed x64 compilation

+++ src/PakLib/PakInterface.cpp
@@ -251,7 +251,7 @@ int PakInterface::FSeek(PFILE *theFile, long theOffset, int theOrigin)
 }
 
 //0x5D8830
-int PakInterface::FTell(PFILE *theFile)
+long PakInterface::FTell(PFILE *theFile)
 {
 	if (theFile->mRecord != NULL)
 		return theFile->mPos;
